requirements.txt
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncpg==0.31.0
bcrypt==5.0.0
click==8.1.8
dotenv==0.9.9
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.124.4
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.49.3
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unicorn==2.1.4
uvicorn==0.38.0
